backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
instructor==1.6.3
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
openai==1.54.3
pillow==11.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
